test(nav-bar): adapt unit tests
